refactor: extract registry domain into orbit-registry [ORB-10302]
