+++ .orbit/adrs/accepted/ADR-0235/adr.yaml
@@ -1,0 +1,28 @@
+schema_version: 2
+id: ADR-0235
+title: Make orbit-registry the singular host/workspace registry domain crate
+status: accepted
+owner: codex
+created_at: 2026-07-18T18:58:37.266035863Z
+accepted_at: 2026-07-18T18:58:42.207312669Z
+last_updated: 2026-07-18T18:58:42.207312669Z
+related_features:
+- host-registry
+- mcp-bridge
+related_tasks:
+- ORB-10302
+tags:
+- crate-boundary
+- host-registry
+- singular-hub
+paths:
+- crates/orbit-registry/**
+- crates/orbit-core/src/host_registry.rs
+- crates/orbit-core/src/routines/host.rs
+- crates/orbit-core/src/workspace_registry.rs
+- crates/orbit-core/src/registry_cache.rs
+- crates/orbit-store/**
+supersedes: []
+legacy_ids: []
+validation_warnings: []
+legacy_validation: none
