+++ crates/orbit-registry/src/host_identity.rs
@@ -1,0 +1,492 @@
+//! Host identity for this Orbit machine [ORB-10247].
+//!
+//! `~/.orbit/host.toml` carries the one genuinely host-local datum: a versioned
+//! [`HostIdentity`] with a stable, generated `machine_id`, an operator-chosen
+//! `host_id` display name, and an operating `mode`. First-time creation lives in
+//! global `orbit init` (implemented by `orbit-core`); a legacy file that carried
+//! only `host_id` (the routines-v1 scheduling pin) is migrated in place, once,
+//! preserving its `host_id` and generating a `machine_id`.
+//!
+//! Loading is strict: after migration an absent, malformed, incomplete, blank,
+//! or future-schema file is a hard error with an actionable message — there is
+//! no silent fallback to the OS hostname (ADR-0227). Routine `hosts:` pinning,
+//! the sweep, and status all resolve through [`HostIdentity::host_id`].
+
+use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::time::{SystemTime, UNIX_EPOCH};
+
+use orbit_common::types::{OrbitError, validate_machine_id};
+use orbit_common::utility::fs::atomic_write_text;
+use serde::Deserialize;
+
+/// File under the global Orbit root carrying the host identity.
+pub const HOST_TOML_FILE: &str = "host.toml";
+
+/// Current on-disk schema version for [`HostIdentity`]. A file whose
+/// `schema_version` exceeds this fails closed (see [`inspect_host_identity`]).
+pub const HOST_IDENTITY_SCHEMA_VERSION: u32 = 1;
+
+const MACHINE_ID_PREFIX: &str = "hm_";
+
+/// Operating mode of this Orbit host.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum HostMode {
+    /// Self-contained host (the standalone default; no registry/hub role).
+    Standalone,
+    /// Coordination hub for a multi-host constellation.
+    Hub,
+    /// Satellite that polls a hub for placed runs.
+    Spoke,
+}
+
+impl HostMode {
+    /// The persisted string form (matches the `mode` values in `host.toml`).
+    pub fn as_str(self) -> &'static str {
+        match self {
+            HostMode::Standalone => "standalone",
+            HostMode::Hub => "hub",
+            HostMode::Spoke => "spoke",
+        }
+    }
+
+    /// Parse an external (CLI / file) mode string, failing closed on anything
+    /// outside the fixed vocabulary.
+    pub fn parse(value: &str) -> Result<Self, OrbitError> {
+        match value.trim() {
+            "standalone" => Ok(HostMode::Standalone),
+            "hub" => Ok(HostMode::Hub),
+            "spoke" => Ok(HostMode::Spoke),
+            other => Err(OrbitError::InvalidInput(format!(
+                "unknown host mode '{other}' (expected 'standalone', 'hub', or 'spoke')"
+            ))),
+        }
+    }
+}
+
+impl std::fmt::Display for HostMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.as_str())
+    }
+}
+
+/// Versioned machine identity persisted in `~/.orbit/host.toml`.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct HostIdentity {
+    /// On-disk schema version; always [`HOST_IDENTITY_SCHEMA_VERSION`] for a
+    /// value produced by this build.
+    pub schema_version: u32,
+    /// Opaque, generated-once, never-reused stable identity (`hm_<hex>`).
+    pub machine_id: String,
+    /// Operator-chosen, renameable display name; matched against routine
+    /// `hosts:` pins.
+    pub host_id: String,
+    /// Host operating mode.
+    pub mode: HostMode,
+}
+
+impl HostIdentity {
+    /// Deterministic `host.toml` rendering — same identity always serializes to
+    /// the same bytes, so a re-migration or repeated init is a no-op write.
+    /// Operator-influenced string values are escaped for a TOML basic string so
+    /// a quote, backslash, or control character cannot corrupt the file.
+    fn to_toml(&self) -> String {
+        format!(
+            "# Machine identity for this Orbit host [ORB-10247]. Created by\n\
+             # `orbit init`; `machine_id` is generated once and never edited.\n\
+             # `host_id` is the operator-chosen display name matched against\n\
+             # routine `hosts:` pins.\n\
+             schema_version = {}\n\
+             machine_id = \"{}\"\n\
+             host_id = \"{}\"\n\
+             mode = \"{}\"\n",
+            self.schema_version,
+            toml_escape_basic(&self.machine_id),
+            toml_escape_basic(&self.host_id),
+            self.mode.as_str(),
+        )
+    }
+}
+
+/// Escape a value for embedding inside a TOML basic (double-quoted) string,
+/// per the TOML spec's basic-string escape set.
+fn toml_escape_basic(value: &str) -> String {
+    let mut escaped = String::with_capacity(value.len());
+    for ch in value.chars() {
+        match ch {
+            '"' => escaped.push_str("\\\""),
+            '\\' => escaped.push_str("\\\\"),
+            '\n' => escaped.push_str("\\n"),
+            '\r' => escaped.push_str("\\r"),
+            '\t' => escaped.push_str("\\t"),
+            '\u{08}' => escaped.push_str("\\b"),
+            '\u{0C}' => escaped.push_str("\\f"),
+            control if (control as u32) < 0x20 || control as u32 == 0x7f => {
+                escaped.push_str(&format!("\\u{:04X}", control as u32));
+            }
+            other => escaped.push(other),
+        }
+    }
+    escaped
+}
+
+/// The three actionable states of `host.toml`. Malformed, incomplete, blank,
+/// and future-schema files are returned as `Err` by [`inspect_host_identity`],
+/// never as a variant here.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum HostIdentityState {
+    /// A complete, current-schema identity.
+    Present(HostIdentity),
+    /// A legacy pre-migration file carrying only `host_id`.
+    Legacy {
+        /// The preserved, non-blank legacy `host_id`.
+        host_id: String,
+    },
+    /// No `host.toml` exists yet.
+    Absent,
+}
+
+impl HostIdentityState {
+    /// The host name this state resolves to for routine pinning, if any.
+    pub fn host_id(&self) -> Option<&str> {
+        match self {
+            HostIdentityState::Present(identity) => Some(&identity.host_id),
+            HostIdentityState::Legacy { host_id } => Some(host_id),
+            HostIdentityState::Absent => None,
+        }
+    }
+}
+
+/// Outcome of [`ensure_host_identity`].
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum HostIdentityOutcome {
+    /// A fresh identity was created from operator input.
+    Created(HostIdentity),
+    /// A legacy `host_id`-only file was migrated to the current schema.
+    Migrated(HostIdentity),
+    /// A complete current-schema identity already existed; nothing was written.
+    Unchanged(HostIdentity),
+}
+
+impl HostIdentityOutcome {
+    /// The resulting identity, regardless of how it was reached.
+    pub fn identity(&self) -> &HostIdentity {
+        match self {
+            HostIdentityOutcome::Created(identity)
+            | HostIdentityOutcome::Migrated(identity)
+            | HostIdentityOutcome::Unchanged(identity) => identity,
+        }
+    }
+}
+
+/// Operator-supplied fields for a first-time identity creation.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct NewHostIdentity {
+    /// Operator-chosen display name.
+    pub host_id: String,
+    /// Operating mode.
+    pub mode: HostMode,
+}
+
+#[derive(Debug, Deserialize)]
+struct RawHostToml {
+    schema_version: Option<u32>,
+    machine_id: Option<String>,
+    host_id: Option<String>,
+    mode: Option<String>,
+}
+
+fn host_toml_path(global_root: &Path) -> PathBuf {
+    global_root.join(HOST_TOML_FILE)
+}
+
+fn non_blank(value: &Option<String>) -> Option<String> {
+    value
+        .as_deref()
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+        .map(str::to_string)
+}
+
+/// Classify `host.toml` without mutating it. Returns the actionable
+/// [`HostIdentityState`] for absent / legacy / present files, and a hard error
+/// for malformed, incomplete, blank, or future-schema files (fail closed —
+/// never rewrites the file).
+pub fn inspect_host_identity(global_root: &Path) -> Result<HostIdentityState, OrbitError> {
+    let path = host_toml_path(global_root);
+    if !path.exists() {
+        return Ok(HostIdentityState::Absent);
+    }
+    let raw_text = std::fs::read_to_string(&path)
+        .map_err(|error| OrbitError::Io(format!("failed to read '{}': {error}", path.display())))?;
+    let parsed: RawHostToml = toml::from_str(&raw_text).map_err(|error| {
+        OrbitError::InvalidInput(format!(
+            "invalid host identity '{}': {error}",
+            path.display()
+        ))
+    })?;
+
+    match parsed.schema_version {
+        None => {
+            // Legacy files carried only `host_id`. A present, non-blank
+            // `host_id` migrates; anything else is unusable.
+            match non_blank(&parsed.host_id) {
+                Some(host_id) => Ok(HostIdentityState::Legacy { host_id }),
+                None => Err(OrbitError::InvalidInput(format!(
+                    "host identity '{}' is incomplete: no schema_version and no host_id; \
+                     run `orbit init` to create one",
+                    path.display()
+                ))),
+            }
+        }
+        Some(version) if version == HOST_IDENTITY_SCHEMA_VERSION => {
+            let machine_id = parsed.machine_id.as_deref().ok_or_else(|| {
+                OrbitError::InvalidInput(format!(
+                    "host identity '{}' is incomplete: missing or blank machine_id",
+                    path.display()
+                ))
+            })?;
+            validate_machine_id(machine_id).map_err(|error| {
+                OrbitError::InvalidInput(format!(
+                    "host identity '{}' has invalid machine_id: {error}",
+                    path.display()
+                ))
+            })?;
+            let host_id = non_blank(&parsed.host_id).ok_or_else(|| {
+                OrbitError::InvalidInput(format!(
+                    "host identity '{}' is incomplete: missing or blank host_id",
+                    path.display()
+                ))
+            })?;
+            let mode = non_blank(&parsed.mode).ok_or_else(|| {
+                OrbitError::InvalidInput(format!(
+                    "host identity '{}' is incomplete: missing or blank mode",
+                    path.display()
+                ))
+            })?;
+            let mode = HostMode::parse(&mode)?;
+            Ok(HostIdentityState::Present(HostIdentity {
+                schema_version: version,
+                machine_id: machine_id.to_string(),
+                host_id,
+                mode,
+            }))
+        }
+        Some(version) if version > HOST_IDENTITY_SCHEMA_VERSION => {
+            Err(OrbitError::InvalidInput(format!(
+                "host identity '{}' has unsupported schema_version {version}; this build \
+                 supports up to {HOST_IDENTITY_SCHEMA_VERSION}. Upgrade Orbit; the file is \
+                 left unchanged",
+                path.display()
+            )))
+        }
+        Some(version) => Err(OrbitError::InvalidInput(format!(
+            "host identity '{}' has invalid schema_version {version}",
+            path.display()
+        ))),
+    }
+}
+
+/// Strictly load a complete, current-schema identity. Absent and legacy files
+/// are errors here (they resolve through `orbit init`); malformed / future
+/// files propagate from [`inspect_host_identity`]. Never falls back to the OS
+/// hostname.
+pub fn load_host_identity(global_root: &Path) -> Result<HostIdentity, OrbitError> {
+    match inspect_host_identity(global_root)? {
+        HostIdentityState::Present(identity) => Ok(identity),
+        HostIdentityState::Legacy { .. } => Err(OrbitError::InvalidInput(format!(
+            "host identity '{}' is a legacy pre-migration file; run `orbit init` to migrate it",
+            host_toml_path(global_root).display()
+        ))),
+        HostIdentityState::Absent => Err(OrbitError::InvalidInput(format!(
+            "no host identity at '{}'; run `orbit init` to create one",
+            host_toml_path(global_root).display()
+        ))),
+    }
+}
+
+/// Ensure a complete, current-schema identity exists, creating or migrating as
+/// needed and returning what happened. `new` supplies the operator's host name
+/// and mode and is invoked **only** when the identity is absent, so callers can
+/// defer prompting until a fresh create is actually required.
+///
+/// Idempotent: a present identity is returned `Unchanged` with no write. A
+/// malformed or future-schema file is never overwritten — the error propagates.
+pub fn ensure_host_identity(
+    global_root: &Path,
+    new: impl FnOnce() -> Result<NewHostIdentity, OrbitError>,
+) -> Result<HostIdentityOutcome, OrbitError> {
+    match inspect_host_identity(global_root)? {
+        HostIdentityState::Present(identity) => Ok(HostIdentityOutcome::Unchanged(identity)),
+        HostIdentityState::Legacy { host_id } => {
+            // Migrate atomically: preserve host_id, generate machine_id, default
+            // to standalone. Rollback leaves the last valid file readable.
+            let identity = HostIdentity {
+                schema_version: HOST_IDENTITY_SCHEMA_VERSION,
+                machine_id: generate_machine_id(),
+                host_id,
+                mode: HostMode::Standalone,
+            };
+            write_host_identity(global_root, &identity)?;
+            Ok(HostIdentityOutcome::Migrated(identity))
+        }
+        HostIdentityState::Absent => {
+            let NewHostIdentity { host_id, mode } = new()?;
+            let host_id = host_id.trim().to_string();
+            if host_id.is_empty() {
+                return Err(OrbitError::InvalidInput(
+                    "host name must not be empty".to_string(),
+                ));
+            }
+            let identity = HostIdentity {
+                schema_version: HOST_IDENTITY_SCHEMA_VERSION,
+                machine_id: generate_machine_id(),
+                host_id,
+                mode,
+            };
+            write_host_identity(global_root, &identity)?;
+            Ok(HostIdentityOutcome::Created(identity))
+        }
+    }
+}
+
+/// Atomically (re)write `host.toml`. The staged-rename write never leaves a
+/// partially overwritten file, so a crash preserves the last valid identity.
+fn write_host_identity(global_root: &Path, identity: &HostIdentity) -> Result<PathBuf, OrbitError> {
+    let text = stage_host_identity_toml(identity)?;
+    write_host_identity_text(global_root, &text)
+}
+
+/// Atomically write already-staged `host.toml` bytes through the crash-safe
+/// atomic-write seam.
+fn write_host_identity_text(global_root: &Path, text: &str) -> Result<PathBuf, OrbitError> {
+    let path = host_toml_path(global_root);
+    atomic_write_text(&path, text).map_err(|error| {
+        OrbitError::Io(format!("failed to write '{}': {error}", path.display()))
+    })?;
+    Ok(path)
+}
+
+/// Render `host.toml` and reparse the staged bytes before any write, confirming
+/// every field round-trips to the intended identity. A quote, backslash, or
+/// control character that would corrupt the file is caught here — before
+/// mutation — rather than producing an unreadable on-disk identity.
+fn stage_host_identity_toml(identity: &HostIdentity) -> Result<String, OrbitError> {
+    let rendered = identity.to_toml();
+    let parsed: RawHostToml = toml::from_str(&rendered).map_err(|error| {
+        OrbitError::InvalidInput(format!(
+            "staged host identity render is not valid TOML: {error}"
+        ))
+    })?;
+    let round_trips = parsed.schema_version == Some(identity.schema_version)
+        && non_blank(&parsed.machine_id).as_deref() == Some(identity.machine_id.as_str())
+        && non_blank(&parsed.host_id).as_deref() == Some(identity.host_id.as_str())
+        && non_blank(&parsed.mode).as_deref() == Some(identity.mode.as_str());
+    if !round_trips {
+        return Err(OrbitError::InvalidInput(
+            "staged host identity render does not round-trip to the intended identity; \
+             refusing to write a corrupt host.toml"
+                .to_string(),
+        ));
+    }
+    Ok(rendered)
+}
+
+/// Rename the current machine's local `host.toml` in place, preserving
+/// `machine_id` and `mode`. The identity must already be `Present`; a legacy or
+/// absent file is a hard error (there is no local identity to rename). The new
+/// render is staged and reparsed before the atomic write, so quotes,
+/// backslashes, and control characters either round-trip safely or fail before
+/// mutation and the last valid file is preserved. Renaming another machine must
+/// never call this — it only ever touches the local file.
+pub fn rename_current_host_identity(
+    global_root: &Path,
+    new_host_id: &str,
+) -> Result<HostIdentity, OrbitError> {
+    rename_current_host_identity_with_writer(global_root, new_host_id, |path, staged| {
+        atomic_write_text(path, staged)
+    })
+}
+
+pub(crate) fn rename_current_host_identity_with_writer<W>(
+    global_root: &Path,
+    new_host_id: &str,
+    writer: W,
+) -> Result<HostIdentity, OrbitError>
+where
+    W: FnOnce(&Path, &str) -> std::io::Result<()>,
+{
+    let current = load_host_identity(global_root)?;
+    let new_host_id = new_host_id.trim();
+    if new_host_id.is_empty() {
+        return Err(OrbitError::InvalidInput(
+            "host name must not be empty".to_string(),
+        ));
+    }
+    let candidate = HostIdentity {
+        schema_version: current.schema_version,
+        machine_id: current.machine_id.clone(),
+        host_id: new_host_id.to_string(),
+        mode: current.mode,
+    };
+    let staged = stage_host_identity_toml(&candidate)?;
+    let path = host_toml_path(global_root);
+    match writer(&path, &staged) {
+        Ok(()) => Ok(candidate),
+        Err(error) => match load_host_identity(global_root) {
+            Ok(observed) if observed == candidate => Err(OrbitError::Io(format!(
+                "host identity write reported an error ({error}), but the complete renamed \
+                 host.toml is now readable; its durability is uncertain"
+            ))),
+            Ok(observed) if observed == current => Err(OrbitError::Io(format!(
+                "host identity write failed ({error}); the previous host.toml is preserved"
+            ))),
+            Ok(observed) => Err(OrbitError::Io(format!(
+                "host identity write failed ({error}); reopening found an unexpected complete \
+                 identity for machine '{}' named '{}', so the local outcome is uncertain",
+                observed.machine_id, observed.host_id
+            ))),
+            Err(reopen) => Err(OrbitError::Io(format!(
+                "host identity write failed ({error}); reopening host.toml to classify \
+                 preservation also failed: {reopen}"
+            ))),
+        },
+    }
+}
+
+/// Best-effort OS hostname, used as the interactive default host name at init.
+/// `None` when the hostname is unavailable or empty.
+pub fn os_hostname() -> Option<String> {
+    hostname::get()
+        .ok()
+        .map(|name| name.to_string_lossy().trim().to_string())
+        .filter(|name| !name.is_empty())
+}
+
+static MACHINE_ID_SEQUENCE: AtomicU64 = AtomicU64::new(0);
+
+/// Generate an opaque, stable `hm_<hex>` machine id. Called exactly once per
+/// machine (at create / migrate); the persisted value is never regenerated.
+fn generate_machine_id() -> String {
+    let nanos = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .map(|d| d.as_nanos() as u64)
+        .unwrap_or(0);
+    let pid = u64::from(std::process::id());
+    let seq = MACHINE_ID_SEQUENCE.fetch_add(1, Ordering::Relaxed);
+    let seed =
+        nanos ^ pid.wrapping_mul(0x9E37_79B9_7F4A_7C15) ^ seq.wrapping_mul(0xD1B5_4A32_D192_ED03);
+    format!("{MACHINE_ID_PREFIX}{:016x}", splitmix64(seed))
+}
+
+/// SplitMix64 finalizer — folds a seed into a well-distributed 64-bit value.
+fn splitmix64(seed: u64) -> u64 {
+    let mut z = seed.wrapping_add(0x9E37_79B9_7F4A_7C15);
+    z = (z ^ (z >> 30)).wrapping_mul(0xBF58_476D_1CE4_E5B9);
+    z = (z ^ (z >> 27)).wrapping_mul(0x94D0_49BB_1331_11EB);
+    z ^ (z >> 31)
+}
+
+#[cfg(test)]
+#[path = "tests/host_identity.rs"]
+mod tests;
