+++ docs/design/mcp-bridge/4_decisions.md
@@ -8,17 +8,18 @@ doc_role: decisions
 type: design
 summary: Accepted ADR log for the coupled MCP Bridge and Host Registry v1 contract.
 tags: [mcp, remote-access, host-registry, bridge]
-paths: ["crates/orbit-mcp/**", "crates/orbit-cli/src/command/mcp/**", "crates/orbit-core/src/command/tool.rs"]
+paths: ["crates/orbit-mcp/**", "crates/orbit-cli/src/command/mcp/**", "crates/orbit-registry/**", "crates/orbit-core/src/command/tool.rs"]
 related_features: [mcp-bridge, host-registry, mcp-session-context, remote-access]
-related_artifacts: [ORB-00424, ORB-10245, ORB-10267, ADR-0226, ADR-0227, ADR-0228, ADR-0229, ADR-0230, ADR-0231, ADR-0232]
+related_artifacts: [ORB-00424, ORB-10245, ORB-10267, ORB-10302, ADR-0226, ADR-0227, ADR-0228, ADR-0229, ADR-0230, ADR-0231, ADR-0232, ADR-0235]
 ---
 
 # Orbit MCP Bridge — Decisions
 
 ADR log for `mcp-bridge`. Entries are append-only and ordered by global ID. The
 Orbit ADR store owns allocation, status, and task links; this log records the
-complete seven-decision v1 contract shared with
-[host-registry](../host-registry/4_decisions.md).
+complete seven-decision v1 behavior contract shared with
+[host-registry](../host-registry/4_decisions.md), plus the crate boundary that
+keeps its registry implementation singular.
 
 ## ADR-0226 — Singular coordination hub, workspace owner, and per-run placement
 
@@ -156,6 +157,29 @@ for its non-Orbit constellation domains.
 - Clients register Orbit and Bridge side by side during migration.
 - Cost: cutover temporarily maintains two registrations and requires deletion of a compatibility layer.
 
+## ADR-0235 — Make orbit-registry the singular host/workspace registry domain crate
+
+**Status:** Accepted · 2026-07 · [ORB-10302] established the coupled implementation boundary.
+
+### Context
+
+The bridge needs one registry-domain API below `orbit-core`, not a second set of
+MCP DTOs and not the retired opaque-byte replication substrate that previously
+occupied `orbit-registry`.
+
+### Decision
+
+Place identity, local workspace roles/catalog, cache semantics, and the
+store-backed service in `orbit-registry`; keep MCP serialization/dispatch in the
+adapter, runtime profile/ship construction in `orbit-core`, persistence in
+`orbit-store`, and shared DTOs in `orbit-common`.
+
+### Consequences
+
+- Future broker and transport work consumes one domain surface without moving SQL or routing policy into it.
+- The retired replicated-writer API cannot reappear as a second authority beside the singular hub.
+- Cost: `orbit-registry` gains a store dependency and is no longer a consumer-agnostic leaf.
+
 ## Task References
 
 - [ORB-00424] — completed design proposal for canonical Orbit MCP and Bridge parity retirement.
@@ -167,5 +191,7 @@ for its non-Orbit constellation domains.
   path-free registry snapshot. C3 proves the real runtime/store action path with no session
   workspace or checkout binding for the enumerated workspace; D3 still owns construction of the
   checkoutless broker plus effective-session `tools/list` omission and `tools/call` denial.
+- [ORB-10302] — moved the coupled registry domain into `orbit-registry` and retained
+  MCP ownership of serialization/dispatch only ([ADR-0235]).
 
 > Resolve any task above with `orbit task show <ID>` or `git log --grep=<ID>`.
